chore: moved tests
