+++ src/helpers/textHelper.test.ts
@@ -1,4 +1,4 @@
-import { truncateText } from "../textHelper";
+import { truncateText } from "./textHelper";
 
 describe("truncateText", () => {
   it("should return the original text if it is shorter than maxChar", () => {
